+++ db/migrate/20230614032215_create_notifications.rb
@@ -1,7 +1,7 @@
 class CreateNotifications < ActiveRecord::Migration[6.1]
   def change
     create_table :notifications do |t|
-      t.references :subject, type: :integer, polymorphic: true
+      t.references :subject, type: :bigint, polymorphic: true
       t.references :end_user, type: :bigint, foreign_key: true
       t.integer :action_type, null: false
